+++ crates/matrix-sdk-crypto/src/machine/mod.rs
@@ -72,9 +72,9 @@ use crate::{
     },
     session_manager::{GroupSessionManager, SessionManager},
     store::{
-        Changes, CryptoStoreWrapper, DeviceChanges, IdentityChanges, IntoCryptoStore, MemoryStore,
-        PendingChanges, Result as StoreResult, RoomKeyInfo, RoomSettings, SecretImportError, Store,
-        StoreCache, StoreTransaction,
+        Changes, CryptoStoreWrapper, DehydratedDeviceKey, DeviceChanges, IdentityChanges,
+        IntoCryptoStore, MemoryStore, PendingChanges, Result as StoreResult, RoomKeyInfo,
+        RoomSettings, SecretImportError, Store, StoreCache, StoreTransaction,
     },
     types::{
         events::{
@@ -2376,6 +2376,33 @@ impl OlmMachine {
     /// Manage dehydrated devices.
     pub fn dehydrated_devices(&self) -> DehydratedDevices {
         DehydratedDevices { inner: self.to_owned() }
+    }
+
+    /// Get the cached dehydrated device pickle key if any.
+    ///
+    /// None if the key was not previously cached (via
+    /// [`Self::save_dehydrated_device_pickle_key`]).
+    ///
+    /// Should be used to periodically rotate the dehydrated device to avoid
+    /// OTK exhaustion and accumulation of to_device messages.
+    pub async fn get_dehydrated_device_pickle_key(
+        &self,
+    ) -> StoreResult<Option<DehydratedDeviceKey>> {
+        self.inner.store.load_dehydrated_device_pickle_key().await
+    }
+
+    /// Store the dehydrated device pickle key in the crypto store.
+    ///
+    /// Use `None` to delete any previously saved pickle key.
+    ///
+    /// This is useful if the client wants to periodically rotate dehydrated
+    /// devices to avoid OTK exhaustion and accumulated to_device problems.
+    pub async fn save_dehydrated_device_pickle_key(
+        &self,
+        dehydrated_device_pickle_key: Option<DehydratedDeviceKey>,
+    ) -> Result<(), CryptoStoreError> {
+        let changes = Changes { dehydrated_device_pickle_key, ..Default::default() };
+        self.inner.store.save_changes(changes).await
     }
 
     /// Get the stored encryption settings for the given room, such as the